+++ backend/baton/src/main/java/touch/baton/config/OauthHttpInterfaceConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-import touch.baton.infra.oauth.github.http.GithubHttpInterface;
+import touch.baton.infra.auth.oauth.github.http.GithubHttpInterface;
 
 import static org.springframework.web.reactive.function.client.support.WebClientAdapter.forClient;
