+++ backend/baton/src/main/java/touch/baton/infra/auth/oauth/github/authcode/GithubAuthCodeRequestUrlProvider.java
@@ -1,11 +1,11 @@
-package touch.baton.infra.oauth.github.authcode;
+package touch.baton.infra.auth.oauth.github.authcode;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
 import touch.baton.domain.oauth.OauthType;
 import touch.baton.domain.oauth.authcode.AuthCodeRequestUrlProvider;
-import touch.baton.infra.oauth.github.GithubOauthConfig;
+import touch.baton.infra.auth.oauth.github.GithubOauthConfig;
 
 @RequiredArgsConstructor
 @Component
