+++ backend/baton/src/main/java/touch/baton/infra/auth/oauth/github/http/GithubHttpInterface.java
@@ -1,12 +1,12 @@
-package touch.baton.infra.oauth.github.http;
+package touch.baton.infra.auth.oauth.github.http;
 
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.service.annotation.GetExchange;
 import org.springframework.web.service.annotation.PostExchange;
-import touch.baton.infra.oauth.github.request.GithubTokenRequest;
-import touch.baton.infra.oauth.github.response.GithubMemberResponse;
-import touch.baton.infra.oauth.github.token.GithubToken;
+import touch.baton.infra.auth.oauth.github.request.GithubTokenRequest;
+import touch.baton.infra.auth.oauth.github.response.GithubMemberResponse;
+import touch.baton.infra.auth.oauth.github.token.GithubToken;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
